# Conflicts:
#   Frontend/my-app/src/Board.js

+++ Backend/controllers/friendController.js
@@ -9,10 +9,10 @@ const db = mysql.createConnection({
 });
 
 exports.getFriends = (request, respond) => {
-    const userID = request.body.userID;
+    const userID = request.params.id;
     console.log(userID);
     var friendList = [];
-    db.query("select * from friendslist where userID = ?;", [userID], (error, result) => {
+    db.query("select friendID from friendslist where userID = ?;", [userID], (error, result) => {
         if(error) {
             console.log(error);
         } else {            
@@ -28,11 +28,9 @@ exports.getFriends = (request, respond) => {
 }
 
 exports.addFriend = (request, respond) => {
-    
     try{
-        const userID = request.body.userID;
+        const userID = request.params.id;
         const friendID = request.body.friendID;
-
         db.query("insert into friendslist(userID, friendID) values(?, ?);", [userID, friendID], (error, result) =>{
 
             //TO-DO
@@ -44,12 +42,14 @@ exports.addFriend = (request, respond) => {
                 console.log("Friend relationship 1");
             }
         })
-    
         db.query("insert into friendslist(userID, friendID) values(?, ?);", [friendID, userID], (error, result) =>{
             if(error) {
                 console.log(error);
             } else {
                 console.log("Friend relationship 2");
+                respond.json({
+                    message: "add friend"
+                })
             }
         })
     }
@@ -61,18 +61,22 @@ exports.addFriend = (request, respond) => {
 exports.deleteFriend = (request, respond) => {
 
     try{
-        const friendID= request.body.friendID;
+        const userID = request.params.id;
+        const friendID = request.body.friendID;
+        db.query("delete from friendslist where friendID = ? and userID =?;", [friendID, userID], (error, result) => {
 
-        db.query("delete from friendslist where friendID = ?;", [friendID], (error, result) => {
+        })
+        db.query("delete from friendslist where friendID = ? and userID =?;", [userID, friendID], (error, result) => {
 
             //TO-DO
             //DELETE AUTHENTICATION (CHECKING IF RESULT IS AN ACTUAL USERNAME)
-            
-
             if(error) {
                 console.log(error);
             } else {
                 console.log("Delete friend");
+                respond.json({
+                    message: "Delete friend"
+                })
             }
         })
     }
